# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CryptoCampus API
  version: 1.0.0
  description: |
    团队 01 的 HTTP 接口唯一机器可读契约。所有业务接口以 /api/v1 为前缀。
    修改字段、类型、错误码或权限前，必须先评审本文件。
...
servers:
  - url: /api/v1
    description: 当前部署
...
paths:
...
  /seals:
    post:
      tags: [Verify]
      operationId: createSeal
      summary: 对文件摘要和时间戳签章
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: [file, private_key]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file, seal_profile, pqc_mode, output_format]
              properties:
                file: { type: string, format: binary }
                seal_profile: { type: string, enum: [personal, department, academic] }
                pqc_mode: { type: boolean }
                output_format:
                  type: string
                  enum: [sidecar, qr, pdf_signature_page]
                  description: pdf_signature_page 为课程进阶能力
      responses:
        "201":
          description: 侧车签章元数据
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Seal" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
  /verifications:
    post:
      tags: [Verify]
      operationId: verifyFile
      summary: 验证文件摘要、签名、证书链、时间戳和吊销状态
      x-required-roles: [guest, student, admin, teacher]
      x-jwt: false
      x-sensitive-fields: [file]
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file, seal]
              properties:
                file: { type: string, format: binary }
                seal: { type: string, format: binary }
      responses:
        "200":
          description: 五步验证结果；验证失败仍返回 200 和 valid=false
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerificationResult" }
        "422":
          $ref: "#/components/responses/ValidationError"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /seals/{seal_id}:
    get:
      tags: [Verify]
      operationId: getSeal
      summary: 通过二维码标识公开获取签章元数据
      x-required-roles: [guest, student, admin, teacher]
      x-jwt: false
      x-sensitive-fields: [source_file]
      security: []
      parameters:
        - $ref: "#/components/parameters/SealId"
      responses:
        "200":
          description: 签章元数据
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Seal" }
        "404": { $ref: "#/components/responses/NotFound" }
  /seals/{seal_id}/sidecar:
    get:
      tags: [Verify]
      operationId: downloadSealSidecar
      summary: 下载侧车签章文件
      x-required-roles: [guest, student, admin, teacher]
      x-jwt: false
      x-sensitive-fields: [source_file]
      security: []
      parameters:
        - $ref: "#/components/parameters/SealId"
      responses:
        "200":
          description: 侧车签章文件
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: cc_refresh
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
...
    SealId:
      name: seal_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: 未认证、凭据错误或会话失效
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: 资源不存在或为防枚举隐藏存在性
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: 字段校验失败
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    PayloadTooLarge:
      description: 文件超过当前操作限制；密信文件的课程上限为 100 MiB
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServiceUnavailable:
      description: 密码引擎或 Provider 暂不可用
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id, details]
      properties:
        code: { type: string, pattern: "^[A-Z][A-Z0-9_]+$" }
        message: { type: string }
        request_id: { type: string }
        details: { type: object, additionalProperties: true }
...
    Seal:
      type: object
      additionalProperties: false
      required: [id, digest_algorithm, digest, signature_algorithm, signature, certificate, timestamp]
      properties:
        id: { type: string, format: uuid }
        digest_algorithm: { const: SM3 }
        digest: { type: string, contentEncoding: base64 }
        signature_algorithm: { type: string, enum: [SM3-with-SM2, ML-DSA-65] }
        signature: { type: string, contentEncoding: base64 }
        certificate: { type: string, contentEncoding: base64 }
        timestamp: { type: string, format: date-time }
    VerificationStep:
      type: object
      additionalProperties: false
      required: [name, passed, message]
      properties:
        name: { type: string, enum: [digest, signature, certificate_chain, timestamp, revocation] }
        passed: { type: boolean }
        message: { type: string }
    VerificationResult:
      type: object
      additionalProperties: false
      required: [valid, steps, record_digest]
      properties:
        valid: { type: boolean }
        steps: { type: array, items: { $ref: "#/components/schemas/VerificationStep" } }
        record_digest: { type: string }
...
security:
  - bearerAuth: []